requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.31.0
stripe==7.0.0
sendgrid==6.11.0
twilio==8.10.0
Flask-Login==0.6.3
oauthlib==3.2.2
PyJWT==2.8.0
Flask-Dance==7.0.0
psycopg2-binary==2.9.9
pgeocode==0.4.1
